deepks : rearrange and add comment
