+++ source/src_lcao/CMakeLists.txt
@@ -62,8 +62,9 @@ list(APPEND objects
 if(ENABLE_DEEPKS)
     list(APPEND objects
         LCAO_descriptor.cpp
-        LCAO_descriptor_new.cpp
         LCAO_descriptor_old.cpp
+        LCAO_descriptor_dV.cpp
+        LCAO_descriptor_io.cpp
         )
 endif()
