+++ source/src_lcao/LCAO_descriptor.cpp
@@ -3,10 +3,16 @@
 
 #include "LCAO_descriptor.h"
 
+//===============================
+//DeePKS Part 1
+//deals with generation of descriptors as well as labels
+//===============================
+
 namespace GlobalC
 {
     LCAO_Descriptor ld;
 }
+
 LCAO_Descriptor::LCAO_Descriptor()
 {
     alpha_index = new ModuleBase::IntArray[1];
@@ -19,7 +25,6 @@ LCAO_Descriptor::LCAO_Descriptor()
     DH_V_delta_y = new double[1];
     DH_V_delta_z = new double[1];
 }
-
 
 LCAO_Descriptor::~LCAO_Descriptor()
 {
@@ -70,10 +75,6 @@ LCAO_Descriptor::~LCAO_Descriptor()
         }
     }
 }
-
-//===============================
-//Part 1. deals with generation of descriptors as well as labels
-//===============================
 
 void LCAO_Descriptor::init(
 	const int lm, // max L for descriptor 
@@ -809,1000 +810,6 @@ void LCAO_Descriptor::del_gdmx(void)
     delete[] this->gdmx;
     delete[] this->gdmy;
     delete[] this->gdmz;
-    return;
-}
-
-//===============================
-//Part 2. deals with application of correction dV to Hamiltonian and force
-//===============================
-
-void LCAO_Descriptor::deepks_pre_scf(const string& model_file)
-{
-    ModuleBase::TITLE("LCAO_Descriptor", "deepks_pre_scf");
-
-	// load the DeePKS model from deep neural network
-    this->load_model(model_file);
-    
-    //initialize the H matrix H_V_delta
-    if(GlobalV::GAMMA_ONLY_LOCAL)
-    {
-        delete[] this->H_V_delta;
-        this->H_V_delta = new double[GlobalC::ParaO.nloc];
-        ModuleBase::GlobalFunc::ZEROS(this->H_V_delta, GlobalC::ParaO.nloc);
-    }
-    else
-    {
-        H_V_delta_k = new std::complex<double>* [GlobalC::kv.nks];
-        for(int ik=0;ik<GlobalC::kv.nks;ik++)
-        {
-            this->H_V_delta_k[ik] = new std::complex<double>[GlobalC::ParaO.nloc];
-            ModuleBase::GlobalFunc::ZEROS(this->H_V_delta_k[ik], GlobalC::ParaO.nloc);
-        }
-    }
-
-    //init gedm**
-    const int pdm_size = (this->lmaxd * 2 + 1) * (this->lmaxd * 2 + 1);
-    this->gedm = new double* [this->inlmax];
-    for (int inl = 0;inl < this->inlmax;inl++)
-    {
-        this->gedm[inl] = new double[pdm_size];
-        ModuleBase::GlobalFunc::ZEROS(this->gedm[inl], pdm_size);
-    }
-    if (GlobalV::FORCE)
-    {
-        //init F_delta
-        F_delta.create(GlobalC::ucell.nat, 3);
-        //init DS_mu_alpha**
-        this->DS_mu_alpha_x = new double* [this->inlmax];
-        this->DS_mu_alpha_y = new double* [this->inlmax];
-        this->DS_mu_alpha_z = new double* [this->inlmax];
-        for (int inl = 0;inl < this->inlmax;inl++)
-        {
-            this->DS_mu_alpha_x[inl] = new double[GlobalV::NLOCAL * (2 * this->lmaxd + 1)];
-            this->DS_mu_alpha_y[inl] = new double[GlobalV::NLOCAL * (2 * this->lmaxd + 1)];
-            this->DS_mu_alpha_z[inl] = new double[GlobalV::NLOCAL * (2 * this->lmaxd + 1)];
-            ModuleBase::GlobalFunc::ZEROS(DS_mu_alpha_x[inl], GlobalV::NLOCAL * (2 * this->lmaxd + 1));
-            ModuleBase::GlobalFunc::ZEROS(DS_mu_alpha_y[inl], GlobalV::NLOCAL * (2 * this->lmaxd + 1));
-            ModuleBase::GlobalFunc::ZEROS(DS_mu_alpha_z[inl], GlobalV::NLOCAL * (2 * this->lmaxd + 1));
-        }
-        //init DH_V_delta*
-        if(GlobalV::GAMMA_ONLY_LOCAL)
-        {
-            delete[] DH_V_delta_x;
-            delete[] DH_V_delta_y;
-            delete[] DH_V_delta_z;
-            this->DH_V_delta_x = new double[GlobalC::ParaO.nloc];
-            this->DH_V_delta_y = new double [GlobalC::ParaO.nloc];
-            this->DH_V_delta_z = new double[GlobalC::ParaO.nloc];
-            ModuleBase::GlobalFunc::ZEROS(DH_V_delta_x, GlobalC::ParaO.nloc);
-            ModuleBase::GlobalFunc::ZEROS(DH_V_delta_y, GlobalC::ParaO.nloc);
-            ModuleBase::GlobalFunc::ZEROS(DH_V_delta_z, GlobalC::ParaO.nloc);
-        }
-        else
-        {
-            DH_V_delta_x_k = new std::complex<double>* [GlobalC::kv.nks];
-            DH_V_delta_y_k = new std::complex<double>* [GlobalC::kv.nks];
-            DH_V_delta_z_k = new std::complex<double>* [GlobalC::kv.nks];
-            for(int ik=0;ik<GlobalC::kv.nks;ik++)
-            {
-                this->DH_V_delta_x_k[ik] = new std::complex<double>[GlobalC::ParaO.nloc];
-                this->DH_V_delta_y_k[ik] = new std::complex<double>[GlobalC::ParaO.nloc];
-                this->DH_V_delta_z_k[ik] = new std::complex<double>[GlobalC::ParaO.nloc];
-                ModuleBase::GlobalFunc::ZEROS(this->DH_V_delta_x_k[ik], GlobalC::ParaO.nloc);
-                ModuleBase::GlobalFunc::ZEROS(this->DH_V_delta_y_k[ik], GlobalC::ParaO.nloc);
-                ModuleBase::GlobalFunc::ZEROS(this->DH_V_delta_z_k[ik], GlobalC::ParaO.nloc);
-            }           
-        }
-    }
-    if(!GlobalV::GAMMA_ONLY_LOCAL)
-    {
-        this->nlm_k.resize(GlobalC::ucell.nat);
-    }
-    return;
-}
-
-//for GAMMA_ONLY, search adjacent atoms from I0
-void LCAO_Descriptor::build_v_delta_alpha(const bool& calc_deri)
-{
-    ModuleBase::TITLE("LCAO_Descriptor", "build_v_delta_alpha");
-    ModuleBase::GlobalFunc::ZEROS(this->H_V_delta, GlobalV::NLOCAL * GlobalV::NLOCAL); //init before calculate
-
-    std::vector<double> Rcut;
-	for(int it1=0; it1<GlobalC::ucell.ntype; ++it1)
-        Rcut.push_back(GlobalC::ORB.Phi[it1].getRcut() + GlobalC::ORB.Alpha[0].getRcut());
-    
-    for (int T0 = 0;T0 < GlobalC::ucell.ntype;++T0)
-    {
-        for (int I0 = 0;I0 < GlobalC::ucell.atoms[T0].na;++I0)
-        {
-            const ModuleBase::Vector3<double> tau0 = GlobalC::ucell.atoms[T0].tau[I0];
-            //Rcut in this function may need to be changed ?! (I think the range of adjacent atoms here should be rcut(phi)+rcut(alpha))
-            GlobalC::GridD.Find_atom(GlobalC::ucell, tau0, T0, I0);
-
-            //adj atom pairs
-            for (int ad1 = 0;ad1 < GlobalC::GridD.getAdjacentNum() + 1;++ad1)
-            {
-                const int T1 = GlobalC::GridD.getType(ad1);
-                const int I1 = GlobalC::GridD.getNatom(ad1);
-				//const int iat1 = GlobalC::ucell.itia2iat(T1, I1);
-                const int start1 = GlobalC::ucell.itiaiw2iwt(T1, I1, 0);
-                const ModuleBase::Vector3<double> tau1 = GlobalC::GridD.getAdjacentTau(ad1);
-                const Atom* atom1 = &GlobalC::ucell.atoms[T1];
-                const int nw1_tot = atom1->nw * GlobalV::NPOL;
-                for (int ad2 = 0;ad2 < GlobalC::GridD.getAdjacentNum() + 1;++ad2)
-                {
-                    const int T2 = GlobalC::GridD.getType(ad2);
-					const int I2 = GlobalC::GridD.getNatom(ad2);
-					const int start2 = GlobalC::ucell.itiaiw2iwt(T2, I2, 0);
-					const ModuleBase::Vector3<double> tau2 = GlobalC::GridD.getAdjacentTau(ad2);
-					const Atom* atom2 = &GlobalC::ucell.atoms[T2];
-                    const int nw2_tot = atom2->nw * GlobalV::NPOL;
-
-                    ModuleBase::Vector3<double> dtau10 = tau1 - tau0;
-                    ModuleBase::Vector3<double> dtau20 = tau2 - tau0;
-                    double distance10 = dtau10.norm() * GlobalC::ucell.lat0;
-                    double distance20 = dtau20.norm() * GlobalC::ucell.lat0;
-                    
-                    if (distance10 < Rcut[T1] && distance20 < Rcut[T2])
-                    {
-                        for (int iw1=0; iw1<nw1_tot; ++iw1)
-						{
-							const int iw1_all = start1 + iw1;
-							const int iw1_local = GlobalC::ParaO.trace_loc_row[iw1_all];
-							if(iw1_local < 0)continue;
-							const int iw1_0 = iw1/GlobalV::NPOL;
-
-							for (int iw2=0; iw2<nw2_tot; ++iw2)
-							{
-								const int iw2_all = start2 + iw2;
-								const int iw2_local = GlobalC::ParaO.trace_loc_col[iw2_all];
-								if(iw2_local < 0)continue;
-								const int iw2_0 = iw2/GlobalV::NPOL;
-
-								double nlm[3];
-								nlm[0] = nlm[1] = nlm[2] = 0.0;
-
-								if(!calc_deri)
-								{
-									GlobalC::UOT.snap_psialpha(
-											nlm, 0, tau1, T1,
-											atom1->iw2l[ iw1_0 ], // L1
-											atom1->iw2m[ iw1_0 ], // m1
-											atom1->iw2n[ iw1_0 ], // N1
-											tau2, T2,
-											atom2->iw2l[ iw2_0 ], // L2
-											atom2->iw2m[ iw2_0 ], // m2
-											atom2->iw2n[ iw2_0 ], // n2
-											GlobalC::ucell.atoms[T0].tau[I0], T0, I0, 
-                                            this->inl_index,
-                                            this->gedm);
-                                    int index=iw2_local*GlobalC::ParaO.nrow+iw1_local;
-                                    this->H_V_delta[index] += nlm[0];
-                                }
-								else  // calculate force
-								{
-									GlobalC::UOT.snap_psialpha(
-											nlm, 1, tau1, T1,
-											atom1->iw2l[ iw1_0 ], // L1
-											atom1->iw2m[ iw1_0 ], // m1
-											atom1->iw2n[ iw1_0 ], // N1
-											tau2, T2,
-											atom2->iw2l[ iw2_0 ], // L2
-											atom2->iw2m[ iw2_0 ], // m2
-											atom2->iw2n[ iw2_0 ], // n2
-											GlobalC::ucell.atoms[T0].tau[I0], T0, I0, 
-                                            this->inl_index,
-                                            this->gedm);
-                                    //for Pulay Force
-                                    //GlobalC::LM.set_force(iw1_all, iw2_all, nlm[0], nlm[1], nlm[2], 'N');
-                                    const int ir = GlobalC::ParaO.trace_loc_row[ iw1_all ];
-                                    const int ic = GlobalC::ParaO.trace_loc_col[ iw2_all ];
-                                    const long index = ir * GlobalC::ParaO.ncol + ic;
-                                    this->DH_V_delta_x[index] += nlm[0];
-                                    this->DH_V_delta_y[index] += nlm[1];
-                                    this->DH_V_delta_z[index] += nlm[2];
-                                }
-							}// end iw2
-						}// end iw1
-					} // end distance
-                }//end ad2
-            }//end ad1
-        }//end I0
-    }//end T0
-
-    /*
-    for(int iw1=0;iw1<GlobalV::NLOCAL;iw1++)
-    {
-        for(int iw2=0;iw2<GlobalV::NLOCAL;iw2++)
-        {
-            GlobalV::ofs_running << H_V_delta[iw1*GlobalV::NLOCAL+iw2] << " ";
-        }
-        GlobalV::ofs_running << std::endl;
-    }
-    */
-    return;
-}
-
-void LCAO_Descriptor::add_v_delta(void)
-{
-    ModuleBase::TITLE("LCAO_DESCRIPTOR", "add_v_delta");
-
-    if (GlobalV::GAMMA_ONLY_LOCAL)
-    {
-        for (int iw1 = 0;iw1 < GlobalV::NLOCAL;++iw1)
-        {
-            for (int iw2 = 0;iw2 < GlobalV::NLOCAL;++iw2)
-            {
-				if (!GlobalC::ParaO.in_this_processor(iw1,iw2))
-				{
-					continue;
-				}
-                
-                int iw1_local=GlobalC::ParaO.trace_loc_row[iw1];
-                int iw2_local=GlobalC::ParaO.trace_loc_col[iw2];
-                int index = iw2_local * GlobalC::ParaO.nrow+ iw1_local;
-                GlobalC::LM.set_HSgamma(iw1, iw2, this->H_V_delta[index], 'L');
-            }
-        }
-    }
-    else
-    {
-		ModuleBase::WARNING_QUIT("add_v_delta","should not be used for multi-k; use add_v_delta_k instead");
-        //call set_HSk, complex Matrix
-    }
-	return;
-}
-
-void LCAO_Descriptor::add_v_delta_k(const int &ik)
-{
-    ModuleBase::TITLE("LCAO_DESCRIPTOR", "add_v_delta_k");
-    ModuleBase::GlobalFunc::ZEROS(this->H_V_delta_k[ik], GlobalC::ParaO.nloc);
-
-    const double Rcut_Alpha = GlobalC::ORB.Alpha[0].getRcut();
-
-    for (int T0 = 0; T0 < GlobalC::ucell.ntype; T0++)
-    {
-		Atom* atom0 = &GlobalC::ucell.atoms[T0]; 
-        for (int I0 =0; I0< atom0->na; I0++)
-        {
-            const int iat = GlobalC::ucell.itia2iat(T0,I0);
-            const ModuleBase::Vector3<double> tau0 = atom0->tau[I0];
-            GlobalC::GridD.Find_atom(GlobalC::ucell, atom0->tau[I0] ,T0, I0);
-
-            for (int ad1=0; ad1<GlobalC::GridD.getAdjacentNum()+1 ; ++ad1)
-            {
-                const int T1 = GlobalC::GridD.getType(ad1);
-                const int I1 = GlobalC::GridD.getNatom(ad1);
-                const int start1 = GlobalC::ucell.itiaiw2iwt(T1, I1, 0);
-                const ModuleBase::Vector3<double> tau1 = GlobalC::GridD.getAdjacentTau(ad1);
-				const Atom* atom1 = &GlobalC::ucell.atoms[T1];
-				const int nw1_tot = atom1->nw*GlobalV::NPOL;
-				const double Rcut_AO1 = GlobalC::ORB.Phi[T1].getRcut();
-
-                ModuleBase::Vector3<double> dR1(GlobalC::GridD.getBox(ad1).x, GlobalC::GridD.getBox(ad1).y, GlobalC::GridD.getBox(ad1).z); 
-
-				for (int ad2=0; ad2 < GlobalC::GridD.getAdjacentNum()+1 ; ad2++)
-				{
-					const int T2 = GlobalC::GridD.getType(ad2);
-					const int I2 = GlobalC::GridD.getNatom(ad2);
-					const int start2 = GlobalC::ucell.itiaiw2iwt(T2, I2, 0);
-					const ModuleBase::Vector3<double> tau2 = GlobalC::GridD.getAdjacentTau(ad2);
-					const Atom* atom2 = &GlobalC::ucell.atoms[T2];
-					const int nw2_tot = atom2->nw*GlobalV::NPOL;
-                    ModuleBase::Vector3<double> dR2(GlobalC::GridD.getBox(ad2).x, GlobalC::GridD.getBox(ad2).y, GlobalC::GridD.getBox(ad2).z);
-					
-					const double Rcut_AO2 = GlobalC::ORB.Phi[T2].getRcut();
-                	const double dist1 = (tau1-tau0).norm() * GlobalC::ucell.lat0;
-                	const double dist2 = (tau2-tau0).norm() * GlobalC::ucell.lat0;
-
-					if (dist1 > Rcut_Alpha + Rcut_AO1
-							|| dist2 > Rcut_Alpha + Rcut_AO2)
-					{
-						continue;
-					}
-
-                    const double arg = ( GlobalC::kv.kvec_d[ik] * (dR2-dR1) ) * ModuleBase::TWO_PI;
-                    const std::complex<double> kphase = std::complex <double> ( cos(arg),  sin(arg) );
-
-					for (int iw1=0; iw1<nw1_tot; ++iw1)
-					{
-						const int iw1_all = start1 + iw1;
-						const int iw1_local = GlobalC::ParaO.trace_loc_row[iw1_all];
-						if(iw1_local < 0)continue;
-						const int iw1_0 = iw1/GlobalV::NPOL;
-						for (int iw2=0; iw2<nw2_tot; ++iw2)
-						{
-							const int iw2_all = start2 + iw2;
-							const int iw2_local = GlobalC::ParaO.trace_loc_col[iw2_all];
-							if(iw2_local < 0)continue;
-							const int iw2_0 = iw2/GlobalV::NPOL;
-
-                            double nlm=0.0;
-                            std::vector<double> nlm1 = this->nlm_k[iat][ad1][iw1_all][0];
-                            std::vector<double> nlm2 = this->nlm_k[iat][ad2][iw2_all][0];
-
-                            assert(nlm1.size()==nlm2.size());
-                            int ib=0;
-                            for (int L0 = 0; L0 <= GlobalC::ORB.Alpha[0].getLmax();++L0)
-                            {
-                                for (int N0 = 0;N0 < GlobalC::ORB.Alpha[0].getNchi(L0);++N0)
-                                {
-                                    const int inl = this->inl_index[T0](I0, L0, N0);
-                                    const int nm = 2*L0+1;
-                                    for (int m1 = 0;m1 < 2 * L0 + 1;++m1)
-                                    {
-                                        for (int m2 = 0; m2 < 2 * L0 + 1; ++m2)
-                                        {
-                                            nlm += this->gedm[inl][m1*nm+m2]*nlm1[ib+m1]*nlm2[ib+m2];
-                                        }
-                                    }
-                                    ib+=(2*L0+1);
-                                }
-                            }
-
-                            assert(ib==nlm1.size());
-                            
-                            int iic;
-
-                            if(GlobalV::KS_SOLVER=="genelpa" || GlobalV::KS_SOLVER=="scalapack_gvx")  // save the matrix as column major format
-                            {
-                                iic=iw1_local+iw2_local*GlobalC::ParaO.nrow;
-                            }
-                            else
-                            {
-                                iic=iw1_local*GlobalC::ParaO.ncol+iw2_local;
-                            }
-                            this->H_V_delta_k[ik][iic] += nlm*kphase;
-                            GlobalC::LM.Hloc2[iic] += nlm * kphase;
-
-						}//iw2
-					}//iw1
-				}//ad2
-			}//ad1
-        }
-    }
-}
-
-//The hellmann-feynmann term in force
-void LCAO_Descriptor::cal_f_delta_hf(const ModuleBase::matrix& dm)
-{
-    ModuleBase::TITLE("LCAO_Descriptor", "cal_f_delta_hf");
-    this->F_delta.zero_out();
-    for (int iat = 0; iat < GlobalC::ucell.nat; ++iat)
-    {
-        const int it = GlobalC::ucell.iat2it[iat];
-        const int ia = GlobalC::ucell.iat2ia[iat];
-        const ModuleBase::Vector3<double> tau0 = GlobalC::ucell.atoms[it].tau[ia];
-		GlobalC::GridD.Find_atom(GlobalC::ucell, GlobalC::ucell.atoms[it].tau[ia] ,it, ia);
-		const double Rcut_Alpha = GlobalC::ORB.Alpha[0].getRcut();
-
-        //FOLLOWING ARE CONTRIBUTIONS FROM
-        //VNL DUE TO PROJECTOR'S DISPLACEMENT
-        for (int ad1 =0 ; ad1 < GlobalC::GridD.getAdjacentNum()+1; ad1++)
-        {
-            const int T1 = GlobalC::GridD.getType (ad1);
-            const Atom* atom1 = &GlobalC::ucell.atoms[T1];
-            const int I1 = GlobalC::GridD.getNatom (ad1);
-            const int start1 = GlobalC::ucell.itiaiw2iwt(T1, I1, 0);
-			const ModuleBase::Vector3<double> tau1 = GlobalC::GridD.getAdjacentTau (ad1);
-			const double Rcut_AO1 = GlobalC::ORB.Phi[T1].getRcut();
-
-            for (int ad2 =0 ; ad2 < GlobalC::GridD.getAdjacentNum()+1; ad2++)
-            {
-                const int T2 = GlobalC::GridD.getType (ad2);
-                const Atom* atom2 = &GlobalC::ucell.atoms[T2];
-                const int I2 = GlobalC::GridD.getNatom (ad2);
-                const int start2 = GlobalC::ucell.itiaiw2iwt(T2, I2, 0);
-                const ModuleBase::Vector3<double> tau2 = GlobalC::GridD.getAdjacentTau (ad2);
-                const double Rcut_AO2 = GlobalC::ORB.Phi[T2].getRcut();
-
-                const double dist1 = (tau1-tau0).norm() * GlobalC::ucell.lat0;
-                const double dist2 = (tau2-tau0).norm() * GlobalC::ucell.lat0;
-
-                if (dist1 > Rcut_Alpha + Rcut_AO1
-                        || dist2 > Rcut_Alpha + Rcut_AO2)
-                {
-                    continue;
-                }
-
-                for (int jj = 0; jj < GlobalC::ucell.atoms[T1].nw; jj++)
-                {
-                    const int iw1_all = start1 + jj;
-                    const int mu = GlobalC::ParaO.trace_loc_row[iw1_all];
-                    if(mu<0) continue;
-                    for (int kk = 0; kk < GlobalC::ucell.atoms[T2].nw; kk++)
-                    {
-                        const int iw2_all = start2 + kk;
-                        const int nu = GlobalC::ParaO.trace_loc_col[iw2_all];
-                        if(nu<0) continue;
-                    
-                        double nlm[3] = {0,0,0};
-                                
-                        GlobalC::UOT.snap_psialpha(
-                            nlm, 1,
-                            tau1, T1,
-                            atom1->iw2l[jj], // L2
-                            atom1->iw2m[jj], // m2
-                            atom1->iw2n[jj], // N2
-                            tau2, T2,
-                            atom2->iw2l[kk], // L1
-                            atom2->iw2m[kk], // m1
-                            atom2->iw2n[kk], // n1
-                            tau0, it, ia, 
-                            this->inl_index,
-                            this->gedm); // mohan  add 2021-05-07
-
-                        double nlm1[3] = {0,0,0};
-
-                        const int index = mu * GlobalC::ParaO.ncol + nu;
-
-                        // HF term is minus, only one projector for each atom force.
-
-                        double sum_dm = 0.0;
-                        //remaining: sum for is
-                        this->F_delta(iat, 0) -= 2 * dm(mu, nu) * nlm[0];
-                        this->F_delta(iat, 1) -= 2 * dm(mu, nu) * nlm[1];
-                        this->F_delta(iat, 2) -= 2 * dm(mu, nu) * nlm[2];
-                        //this->F_delta(iat, 0) -= 4 * dm(mu, nu) * nlm[0];   //2 for v_delta(not calculated togethor), 2 for e_delta
-                        //this->F_delta(iat, 1) -= 4 * dm(mu, nu) * nlm[1];
-                        //this->F_delta(iat, 2) -= 4 * dm(mu, nu) * nlm[2];
-                    }//!kk
-                }//!ad2
-            }//!jj
-        }//!ad1
-    }//!iat
-    return;
-}
-
-//the pulay term in force
-void LCAO_Descriptor::cal_f_delta_pulay(const ModuleBase::matrix& dm)
-{
-    ModuleBase::TITLE("LCAO_Descriptor", "cal_f_delta_pulay");
-    //this->F_delta.zero_out();
-    this->build_v_delta_alpha_new(1);
-    //this->build_v_delta_mu(1);    //, if multi-k
-    for (int i = 0;i < GlobalV::NLOCAL;++i)   //col, diff
-    {
-        const int iat = GlobalC::ucell.iwt2iat[i];//the atom whose force being calculated
-        for (int j = 0;j < GlobalV::NLOCAL;++j)   //row
-        {
-            const int mu = GlobalC::ParaO.trace_loc_row[j];
-            const int nu = GlobalC::ParaO.trace_loc_col[i];
-            if (mu >= 0 && nu >= 0)
-            {
-                const int index = mu * GlobalC::ParaO.ncol + nu;
-                this->F_delta(iat, 0) += 2 * dm(nu, mu) * this->DH_V_delta_x[index];
-                this->F_delta(iat, 1) += 2 * dm(nu, mu) * this->DH_V_delta_y[index];
-                this->F_delta(iat, 2) += 2 * dm(nu, mu) * this->DH_V_delta_z[index];
-            }
-        }
-    }
-    return;
-}
-
-void LCAO_Descriptor::cal_e_delta_band(const std::vector<ModuleBase::matrix> &dm)
-{
-    ModuleBase::TITLE("LCAO_Descriptor", "cal_e_delta_band");
-    this->e_delta_band = 0;
-    for (int i = 0; i < GlobalV::NLOCAL; ++i)
-    {
-        for (int j = 0; j < GlobalV::NLOCAL; ++j)
-        {
-            const int mu = GlobalC::ParaO.trace_loc_row[j];
-            const int nu = GlobalC::ParaO.trace_loc_col[i];
-            
-            if (mu >= 0 && nu >= 0)
-            {                
-                const int index=nu*GlobalC::ParaO.nrow+mu;
-                for (int is = 0; is < GlobalV::NSPIN; ++is)
-                {
-                    this->e_delta_band += dm[is](nu, mu) * this->H_V_delta[index];
-                }
-            }
-        }
-    }
-    Parallel_Reduce::reduce_double_all(this->e_delta_band);
-    return;
-}
-
-void LCAO_Descriptor::cal_e_delta_band_k(const std::vector<ModuleBase::ComplexMatrix> &dm)
-{
-    ModuleBase::TITLE("LCAO_Descriptor", "cal_e_delta_band");
-    std::complex<double> e_delta_band_k=std::complex<double>(0.0,0.0);
-    for (int i = 0; i < GlobalV::NLOCAL; ++i)
-    {
-        for (int j = 0; j < GlobalV::NLOCAL; ++j)
-        {
-            const int mu = GlobalC::ParaO.trace_loc_row[j];
-            const int nu = GlobalC::ParaO.trace_loc_col[i];
-            
-            if (mu >= 0 && nu >= 0)
-            {                
-                int iic;
-                if(GlobalV::KS_SOLVER=="genelpa" || GlobalV::KS_SOLVER=="scalapack_gvx")  // save the matrix as column major format
-                {
-                    iic=mu+nu*GlobalC::ParaO.nrow;
-                }
-                else
-                {
-                    iic=mu*GlobalC::ParaO.ncol+nu;
-                }
-                for(int ik=0;ik<GlobalC::kv.nks;ik++)
-                {
-                    e_delta_band_k += dm[ik](nu, mu) * this->H_V_delta_k[ik][iic];
-                }
-            }
-        }
-    }
-    Parallel_Reduce::reduce_complex_double_all(e_delta_band_k);
-    if(e_delta_band_k.imag()>1e-12)
-    {
-        GlobalV::ofs_running << "e_delta_band_k : " << e_delta_band_k << std::endl;
-        ModuleBase::WARNING_QUIT("e_delta_band_k","energy should be real!");
-    }
-    this->e_delta_band = e_delta_band_k.real();
-    return;
-}
-
-//============================
-//Part 3. deals with io as well as interface with libtorch
-//============================
-
-//calculates descriptors from projected density matrices
-void LCAO_Descriptor::cal_descriptor_tensor(void)
-{
-    ModuleBase::TITLE("LCAO_Descriptor", "cal_descriptor_tensor");
-    //init pdm_tensor and d_tensor
-    torch::Tensor tmp;
-
-    //if pdm_tensor and d_tensor is not empty, clear it !!
-    if (!this->d_tensor.empty())
-    {
-        this->d_tensor.erase(this->d_tensor.begin(), this->d_tensor.end());
-    }
-    if (!this->pdm_tensor.empty())
-    {
-        this->pdm_tensor.erase(this->pdm_tensor.begin(), this->pdm_tensor.end());
-    }
-
-    for (int inl = 0;inl < this->inlmax;++inl)
-    {
-        int nm = 2 * inl_l[inl] + 1;
-        tmp = torch::ones({ nm, nm }, torch::TensorOptions().dtype(torch::kFloat64));
-        for (int m1 = 0;m1 < nm;++m1)
-        {
-            for (int m2 = 0;m2 < nm;++m2)
-            {
-                tmp.index_put_({m1, m2}, this->pdm[inl][m1 * nm + m2]);
-            }
-        }
-        //torch::Tensor tmp = torch::from_blob(this->pdm[inl], { nm, nm }, torch::requires_grad());
-        tmp.requires_grad_(true);
-        this->pdm_tensor.push_back(tmp);
-        this->d_tensor.push_back(torch::ones({ nm }, torch::requires_grad(true)));
-    }
-
-    //cal d_tensor
-    for (int inl = 0;inl < inlmax;++inl)
-    {
-        torch::Tensor vd;
-        std::tuple<torch::Tensor, torch::Tensor> d_v(this->d_tensor[inl], vd);
-        //d_v = torch::symeig(pdm_tensor[inl], /*eigenvalues=*/true, /*upper=*/true);
-        d_v = torch::linalg::eigh(pdm_tensor[inl], /*uplo*/"U");
-        d_tensor[inl] = std::get<0>(d_v);
-    }
-    return;
-}
-
-//calculates gradient of descriptors from gradient of projected density matrices
-void LCAO_Descriptor::cal_gvx(const ModuleBase::matrix &dm)
-{
-    ModuleBase::TITLE("LCAO_Descriptor","cal_gvx");
-    //preconditions
-    this->cal_gvdm();
-
-    this->build_S_descriptor(1);
-    this->init_gdmx();
-    this->cal_gdmx(dm); //checked
-
-    //make gdmx as tensor
-    int nlmax = this->inlmax/GlobalC::ucell.nat;
-    for (int nl=0;nl<nlmax;++nl)
-    {
-        std::vector<torch::Tensor> bmmv;
-        for (int ibt=0;ibt<GlobalC::ucell.nat;++ibt)
-        {
-            std::vector<torch::Tensor> xmmv;
-            for (int i=0;i<3;++i)
-            {
-                std::vector<torch::Tensor> ammv;
-                for (int iat=0; iat<GlobalC::ucell.nat; ++iat)
-                {
-                    int inl = iat*nlmax + nl;
-                    int nm = 2*this->inl_l[inl]+1;
-                    std::vector<double> mmv;
-                    for (int m1=0;m1<nm;++m1)
-                    {
-                        for(int m2=0;m2<nm;++m2)
-                        {
-                            if(i==0) mmv.push_back(this->gdmx[ibt][inl][m1*nm+m2]);
-                            if(i==1) mmv.push_back(this->gdmy[ibt][inl][m1*nm+m2]);
-                            if(i==2) mmv.push_back(this->gdmz[ibt][inl][m1*nm+m2]);
-                        }
-                    }//nm^2
-                    torch::Tensor mm = torch::tensor(mmv, torch::TensorOptions().dtype(torch::kFloat64) ).reshape({nm, nm});    //nm*nm
-                    ammv.push_back(mm);
-                }
-                torch::Tensor amm = torch::stack(ammv, 0);  //nat*nm*nm
-                xmmv.push_back(amm);
-            }
-            torch::Tensor bmm = torch::stack(xmmv, 0);  //3*nat*nm*nm
-            bmmv.push_back(bmm); 
-        }
-        this->gdmr_vector.push_back(torch::stack(bmmv, 0)); //nbt*3*nat*nm*nm
-    }
-    assert(this->gdmr_vector.size()==nlmax);
-
-    std::cout<<"gdmr-ok"<<std::endl;
-    std::cout << nlmax <<" " << this->gdmr_vector.size()<<" "<<this->gevdm_vector.size()<<std::endl;
-    //einsum for each inl: 
-    std::vector<torch::Tensor> gvx_vector;
-    for (int nl = 0;nl<nlmax;++nl)
-    {
-        gvx_vector.push_back(at::einsum("bxamn, avmn->bxav", {this->gdmr_vector[nl], this->gevdm_vector[nl]}));
-    }//
-    
-    // cat nv-> \sum_nl(nv) = \sum_nl(nm_nl)=des_per_atom
-    this->gvx_tensor = torch::cat(gvx_vector, -1);
-
-    assert(this->gvx_tensor.size(0) == GlobalC::ucell.nat);
-    assert(this->gvx_tensor.size(1) == 3);
-    assert(this->gvx_tensor.size(2) == GlobalC::ucell.nat);
-    assert(this->gvx_tensor.size(3) == this->des_per_atom);
-
-    return;
-}
-
-void LCAO_Descriptor::load_model(const string& model_file)
-{
-    ModuleBase::TITLE("LCAO_Descriptor", "load_model");
-
-    try
-	{
-        this->module = torch::jit::load(model_file);
-    }
-    catch (const c10::Error& e)
-
-	{
-        std::cerr << "error loading the model" << std::endl;
-        return;
-    }
-	return;
-}
-
-//obtain from the machine learning model dE_delta/dDescriptor
-void LCAO_Descriptor::cal_gedm(const ModuleBase::matrix &dm)
-{
-    //using this->pdm_tensor
-    ModuleBase::TITLE("LCAO_Descriptor", "cal_gedm");
-    //-----prepare for autograd---------
-    this->cal_projected_DM(dm);
-    this->cal_descriptor();
-    this->cal_descriptor_tensor();  //use torch::linalg::eigh
-    //-----prepared-----------------------
-    //forward
-    std::vector<torch::jit::IValue> inputs;
-    //input_dim:(natom, des_per_atom)
-    inputs.push_back(torch::cat(this->d_tensor, /*dim=*/0).reshape({ GlobalC::ucell.nat, this->des_per_atom }));
-    std::vector<torch::Tensor> ec;
-    ec.push_back(module.forward(inputs).toTensor());    //Hartree
-    this->E_delta = ec[0].item().toDouble() * 2;//Ry; *2 is for Hartree to Ry
-    
-    //cal gedm
-    std::vector<torch::Tensor> gedm_shell;
-    gedm_shell.push_back(torch::ones_like(ec[0]));
-    this->gedm_tensor = torch::autograd::grad(ec, this->pdm_tensor, gedm_shell, /*retain_grad=*/true, /*create_graph=*/false, /*allow_unused=*/true);
-
-    //gedm_tensor(Hartree) to gedm(Ry)
-    for (int inl = 0;inl < inlmax;++inl)
-    {
-        int nm = 2 * inl_l[inl] + 1;
-        for (int m1 = 0;m1 < nm;++m1)
-        {
-            for (int m2 = 0;m2 < nm;++m2)
-            {
-                int index = m1 * nm + m2;
-                //*2 is for Hartree to Ry
-                this->gedm[inl][index] = this->gedm_tensor[inl].index({ m1,m2 }).item().toDouble() * 2;
-            }
-        }
-    }
-    return;
-}
-
-//dE_delta/dDescriptor * dDescriptor / dprojected density matrix
-void LCAO_Descriptor::cal_gvdm()
-{
-    ModuleBase::TITLE("LCAO_Descriptor", "cal_gvdm");
-    //cal gevdm(d(EigenValue(D))/dD)
-    int nlmax = inlmax/GlobalC::ucell.nat;
-    for (int nl=0;nl<nlmax;++nl)
-    {
-        std::vector<torch::Tensor> avmmv;
-        for (int iat = 0;iat<GlobalC::ucell.nat;++iat)
-        {
-            int inl = iat*nlmax+nl;
-            int nm = 2*this->inl_l[inl]+1;
-            //repeat each block for nm times in an additional dimension
-            torch::Tensor tmp_x = this->pdm_tensor[inl].reshape({nm, nm}).unsqueeze(0).repeat({nm, 1, 1});
-            //torch::Tensor tmp_y = std::get<0>(torch::symeig(tmp_x, true));
-            torch::Tensor tmp_y = std::get<0>(torch::linalg::eigh(tmp_x, "U"));
-            torch::Tensor tmp_yshell = torch::eye(nm, torch::TensorOptions().dtype(torch::kFloat64));
-            std::vector<torch::Tensor> tmp_rpt;     //repeated-pdm-tensor (x)
-            std::vector<torch::Tensor> tmp_rdt; //repeated-d-tensor (y)
-            std::vector<torch::Tensor> tmp_gst; //gvx-shell
-            tmp_rpt.push_back(tmp_x);
-            tmp_rdt.push_back(tmp_y);
-            tmp_gst.push_back(tmp_yshell);
-            std::vector<torch::Tensor> tmp_res;
-            tmp_res = torch::autograd::grad(tmp_rdt, tmp_rpt, tmp_gst, false, false, /*allow_unused*/true); //nm(v)**nm*nm
-            avmmv.push_back(tmp_res[0]);
-        }
-        torch::Tensor avmm = torch::stack(avmmv, 0); //nat*nv**nm*nm
-        this->gevdm_vector.push_back(avmm);
-    }
-    assert(this->gevdm_vector.size() == nlmax);
-    return;
-}
-
-void LCAO_Descriptor::print_H_V_delta(void)
-{
-    ModuleBase::TITLE("LCAO_Descriptor", "print_H_V_delta");
-
-    ofstream ofs;
-    stringstream ss;
-
-    // the parameter 'winput::spillage_outdir' is read from INPUTw.
-    ss << winput::spillage_outdir << "/"
-       << "H_V_delta.dat";
-
-    if (GlobalV::MY_RANK == 0)
-    {
-        ofs.open(ss.str().c_str());
-    }
-
-    ofs << "E_delta(Ry) from deepks model: " << this->E_delta << std::endl;
-    ofs << "E_delta(eV) from deepks model: " << this->E_delta * ModuleBase::Ry_to_eV << std::endl;
-    ofs << "H_delta(Ry)(gamma only)) from deepks model: " << std::endl;
-
-    for (int i = 0;i < GlobalV::NLOCAL;++i)
-    {
-        for (int j = 0;j < GlobalV::NLOCAL;++j)
-        {
-            ofs<< std::setw(12)<< this->H_V_delta[i * GlobalV::NLOCAL + j] << " ";
-        }
-        ofs << std::endl;
-    }
-    ofs << "H_delta(eV)(gamma only)) from deepks model: " << std::endl;
-
-    for (int i = 0;i < GlobalV::NLOCAL;++i)
-    {
-        for (int j = 0;j < GlobalV::NLOCAL;++j)
-        {
-            ofs<< std::setw(12)<< this->H_V_delta[i * GlobalV::NLOCAL + j] *ModuleBase::Ry_to_eV<< " ";
-        }
-        ofs << std::endl;
-    }
-
-    GlobalV::ofs_running << " H_delta has been printed to " << ss.str() << std::endl;
-    return;
-}
-
-
-void LCAO_Descriptor::print_F_delta(const string& fname)
-{
-    ModuleBase::TITLE("LCAO_Descriptor", "print_F_delta");
-
-    ofstream ofs;
-    stringstream ss;
-    // the parameter 'winput::spillage_outdir' is read from INPUTw.
-    ss << winput::spillage_outdir << "/"<< fname ;
-
-    if (GlobalV::MY_RANK == 0)
-    {
-        ofs.open(ss.str().c_str());
-    }
-
-    ofs << "F_delta(Hatree/Bohr) from deepks model: " << std::endl;
-    ofs << std::setw(12) << "type" << std::setw(12) << "atom" << std::setw(15) << "dF_x" << std::setw(15) << "dF_y" << std::setw(15) << "dF_z" << std::endl;
-
-    for (int it = 0;it < GlobalC::ucell.ntype;++it)
-    {
-        for (int ia = 0;ia < GlobalC::ucell.atoms[it].na;++ia)
-        {
-            int iat = GlobalC::ucell.itia2iat(it, ia);
-            ofs << std::setw(12) << GlobalC::ucell.atoms[it].label << std::setw(12) << ia
-                << std::setw(15) << this->F_delta(iat, 0) / 2 << std::setw(15) << this->F_delta(iat, 1) / 2
-                << std::setw(15) << this->F_delta(iat, 2) / 2 << std::endl;
-        }
-    }
-
-    ofs << "F_delta(eV/Angstrom) from deepks model: " << std::endl;
-    ofs << std::setw(12) << "type" << std::setw(12) << "atom" << std::setw(15) << "dF_x" << std::setw(15) << "dF_y" << std::setw(15) << "dF_z" << std::endl;
-
-    for (int it = 0;it < GlobalC::ucell.ntype;++it)
-    {
-        for (int ia = 0;ia < GlobalC::ucell.atoms[it].na;++ia)
-        {
-            int iat = GlobalC::ucell.itia2iat(it, ia);
-            ofs << std::setw(12) << GlobalC::ucell.atoms[it].label << std::setw(12)
-                << ia << std::setw(15) << this->F_delta(iat, 0) * ModuleBase::Ry_to_eV/ModuleBase::BOHR_TO_A
-                << std::setw(15) << this->F_delta(iat, 1) * ModuleBase::Ry_to_eV/ModuleBase::BOHR_TO_A
-                << std::setw(15) << this->F_delta(iat, 2) * ModuleBase::Ry_to_eV/ModuleBase::BOHR_TO_A << std::endl;
-        }
-    }
-
-    GlobalV::ofs_running << " F_delta has been printed to " << ss.str() << std::endl;
-    ofs.close();
-
-    /*
-    //============for test: double check of 2 methods=============== 
-    //1. same as old method
-    ModuleBase::matrix F_delta_old;
-    F_delta_old.create(GlobalC::ucell.nat, 3);
-    F_delta_old.zero_out();
-    for (int inl = 0;inl < this->inlmax;++inl)
-    {
-        int nm = 2 * inl_l[inl] + 1;
-        for (int m1 = 0;m1 < nm;++m1)
-        {
-            for (int m2 = 0;m2 < nm;++m2)
-            {
-                for (int mu = 0;mu < GlobalV::NLOCAL;++mu)
-                {
-                    int iat = GlobalC::ucell.iwt2iat[mu];
-                    for (int nu = 0;nu < GlobalV::NLOCAL;++nu)
-                    {
-                        F_delta_old(iat, 0) += 2*LOC.wfc_dm_2d.dm_gamma[0](mu, nu)*gedm[inl][m1 * nm + m2] * this->DS_mu_alpha_x[inl][m1 * GlobalV::NLOCAL + mu] * this->S_mu_alpha[inl][m2 * GlobalV::NLOCAL + nu];
-                        F_delta_old(iat, 1) += 2*LOC.wfc_dm_2d.dm_gamma[0](mu, nu)*gedm[inl][m1 * nm + m2] * this->DS_mu_alpha_y[inl][m1 * GlobalV::NLOCAL + mu] * this->S_mu_alpha[inl][m2 * GlobalV::NLOCAL + nu];
-                        F_delta_old(iat,2) += 2*LOC.wfc_dm_2d.dm_gamma[0](mu, nu)*gedm[inl][m1 * nm + m2] * this->DS_mu_alpha_z[inl][m1 * GlobalV::NLOCAL + mu] * this->S_mu_alpha[inl][m2 * GlobalV::NLOCAL + nu];
-                    }
-                }
-            }
-        }
-    }
-    //print F_new
-    stringstream ss1;
-    ss1 << winput::spillage_outdir << "/"
-       << "F_delta_old.dat";
-    if (GlobalV::MY_RANK == 0)
-    {
-        ofs.open(ss1.str().c_str());
-    }
-    for (int iat = 0;iat < GlobalC::ucell.nat;++iat)
-    {
-        for (int i = 0;i < 3;++i)
-        {
-            ofs<< std::setw(8)<< F_delta_old(iat,i)/2<< " ";
-        }
-        ofs << std::endl;
-    }
-    ofs.close();
-
-    //2. same as new method
-    ModuleBase::matrix F_delta_new;
-    F_delta_new.create(GlobalC::ucell.nat, 3);
-    F_delta_new.zero_out();
-    for (int inl = 0;inl < this->inlmax;++inl)
-    {
-        int nm = 2 * inl_l[inl] + 1;
-        for (int m1 = 0;m1 < nm;++m1)
-        {
-            for (int m2 = 0;m2 < nm;++m2)
-            {
-                for (int mu = 0;mu < GlobalV::NLOCAL;++mu)
-                {
-                    for (int nu = 0;nu < GlobalV::NLOCAL;++nu)
-                    {
-                        int iat = GlobalC::ucell.iwt2iat[nu];
-                        F_delta_new(iat, 0) += 2 * LOC.wfc_dm_2d.dm_gamma[0](mu, nu) * gedm[inl][m1 * nm + m2] * this->DS_mu_alpha_x[inl][m1 * GlobalV::NLOCAL + mu] * this->S_mu_alpha[inl][m2 * GlobalV::NLOCAL + nu];
-                        F_delta_new(iat, 1) += 2*LOC.wfc_dm_2d.dm_gamma[0](mu, nu)*gedm[inl][m1 * nm + m2] * this->DS_mu_alpha_y[inl][m1 * GlobalV::NLOCAL + mu] * this->S_mu_alpha[inl][m2 * GlobalV::NLOCAL + nu];
-                        F_delta_new(iat,2) += 2*LOC.wfc_dm_2d.dm_gamma[0](mu, nu)*gedm[inl][m1 * nm + m2] * this->DS_mu_alpha_z[inl][m1 * GlobalV::NLOCAL + mu] * this->S_mu_alpha[inl][m2 * GlobalV::NLOCAL + nu];
-                    }
-                }
-            }
-        }
-    }
-    //print F_new
-    stringstream ss2;
-    ss2 << winput::spillage_outdir << "/"
-       << "F_delta_new.dat";
-    if (GlobalV::MY_RANK == 0)
-    {
-        ofs.open(ss2.str().c_str());
-    }
-    for (int iat = 0;iat < GlobalC::ucell.nat;++iat)
-    {
-        for (int i = 0;i < 3;++i)
-        {
-            ofs<< std::setw(8)<< F_delta_new(iat,i)/2<< " ";
-        }
-        ofs << std::endl;
-    }
-    ofs.close();
-*/
-    return;
-}
-
-
-void LCAO_Descriptor::save_npy_d(void)
-{
-    ModuleBase::TITLE("LCAO_Descriptor", "save_npy_d");
-    //save descriptor in .npy format
-    vector<double> npy_des;
-    for (int i = 0;i < this->n_descriptor;++i)
-    {
-        npy_des.push_back(this->d[i]);
-    }
-    const long unsigned dshape[] = {(long unsigned) GlobalC::ucell.nat, (long unsigned) this->des_per_atom };
-    if (GlobalV::MY_RANK == 0)
-    {
-        npy::SaveArrayAsNumpy("dm_eig.npy", false, 2, dshape, npy_des);
-    }
-    return;
-}
-
-
-void LCAO_Descriptor::save_npy_e(const double &e, const std::string &e_file)
-{
-    ModuleBase::TITLE("LCAO_Descriptor", "save_npy_e");
-    //save e_base
-    const long unsigned eshape[] = { 1 };
-    vector<double> npy_e;
-    npy_e.push_back(e);
-    npy::SaveArrayAsNumpy(e_file, false, 1, eshape, npy_e);
-    return;
-}
-
-void LCAO_Descriptor::save_npy_f(const ModuleBase::matrix &f, const std::string &f_file)
-{
-    ModuleBase::TITLE("LCAO_Descriptor", "save_npy_f");
-    //save f_base
-    //caution: unit: Rydberg/Bohr
-    const long unsigned fshape[] = {(long unsigned) GlobalC::ucell.nat, 3 };
-    vector<double> npy_f;
-    for (int iat = 0;iat < GlobalC::ucell.nat;++iat)
-    {
-        for (int i = 0;i < 3;i++)
-        {
-            npy_f.push_back(f(iat, i));
-        }
-    }
-    npy::SaveArrayAsNumpy(f_file, false, 2, fshape, npy_f);
-    return;
-}
-
-void LCAO_Descriptor::save_npy_gvx(void)
-{
-    ModuleBase::TITLE("LCAO_Descriptor", "save_npy_gvx");
-    //save grad_vx.npy (when  force label is in use)
-    //unit: /Bohr
-    const long unsigned gshape[] = {(long unsigned) GlobalC::ucell.nat, 3, GlobalC::ucell.nat, this->des_per_atom};
-    vector<double> npy_gvx;
-    for (int ibt = 0;ibt < GlobalC::ucell.nat;++ibt)
-    {
-        for (int i = 0;i < 3;i++)
-        {
-            for (int iat = 0;iat < GlobalC::ucell.nat;++iat)
-            {
-                for(int p=0;p<this->des_per_atom;++p)
-                {
-                    npy_gvx.push_back(this->gvx_tensor.index({ ibt, i, iat, p }).item().toDouble());
-                }
-            }
-        }
-    }
-    npy::SaveArrayAsNumpy("grad_vx.npy", false, 4, gshape, npy_gvx);
     return;
 }
