+++ source/src_lcao/LCAO_descriptor.h
@@ -43,10 +43,17 @@
 class LCAO_Descriptor
 {
 
-//-------------------
-// public functions
-//-------------------
+//===============================
+//DeePKS Part 1
+//deals with generation of descriptors as well as labels
+//===============================
 public:
+
+//===============================
+//DeePKS Part 1
+//deals with generation of descriptors as well as labels
+//realized in LCAO_descriptor.cpp
+//===============================
 
     explicit LCAO_Descriptor();
     ~LCAO_Descriptor();
@@ -57,30 +64,57 @@ public:
         const int nm/**< [in] max orbital number with the same 'L', for each 'L'*/,
         const int tot_inl/**< [in] total number of radial orbitals (sum of atoms 'I', angular number 'L' and orbital number 'N') */);
 
-	/// calculate\f$S_{\alpha, \mu} = \langle\alpha|\phi_\mu\rangle \f$ overlap between lcao basis Phi and descriptor basis Alpha
+	/// calculate overlap between lcao basis Phi and descriptor basis Alpha
+    /// <alpha | chi>
     void build_S_descriptor(const bool &calc_deri/**< [in] 0 for \f$\langle\phi|\alpha\rangle\f$, 1 for \f$\langle\frac{d\phi}{dR}|\alpha\rangle\f$*/);
 
+    //S_alpha_mu * DM  * S_nu_beta
+    ///calculate projected density matrix:
+    ///pdm = sum_i,occ <phi_i|alpha1><alpha2|phi_k>
+    void cal_projected_DM(const ModuleBase::matrix& dm/**< [in] density matrix*/);
 
+    ///EIGENVALUE of pdm in block of I_n_l
+    void cal_descriptor(void);
+
+//===============================
+//DeePKS Part 2
+//deals with application of correction dV to Hamiltonian and force
+//realized in LCAO_descriptor_dV.cpp
+//===============================
 	/// 1. Load DeePKS model
     /// 2. Initialize the deltaV Hamiltonian matrix 
     /// 3. If FORCE, initialize the matrces for force
     void deepks_pre_scf(const std::string& model_file/**< [in] path of a traced model file, provided by deepks-kit*/);
-
-    //S_alpha_mu * DM  * S_nu_beta
-    ///calculate projected density matrix:
-    ///\f[D^ I_{ nlmm'} = \sum_{i}\sum_{\mu, \nu}\langle\alpha^I_{nlm}|\phi_\mu\rangle c_{i,\mu}c_{i,\nu} \langle\phi_\nu|\alpha^I_{nlm' }\rangle\f]
-    void cal_projected_DM(const ModuleBase::matrix& dm/**< [in] density matrix*/);
     
-    ///EIGENVALUE of pdm in block of I_n_l
-    void cal_descriptor(void);
+    ///add dV to the Hamiltonian matrix
+    void add_v_delta(void);
+    void add_v_delta_k(const int &ik);
     
-    ///compute the descriptor for each atom
-    void cal_dm_as_descriptor(const ModuleBase::matrix& dm/**< [in] density matrix*/); // mohan add 2021-08-04
+    //calculates sum_(L0,M0) alpha<psi_i|alpha><alpha|psi_j>
+    void build_v_delta_alpha_new(const bool& cal_deri/**< [in] 0 for 3-center intergration, 1 for its derivation*/);
 
-    ///calculate \f$\frac{dE_\delta}{dD^I_{nlmm'}}\f$
+    //Hellman-Feymann term in force
+    void cal_f_delta_hf_new(const ModuleBase::matrix& dm/**< [in] density matrix*/, const bool isstress, ModuleBase::matrix& svnl_dalpha);
+    void cal_f_delta_hf_k_new(const std::vector<ModuleBase::ComplexMatrix>& dm/**<[in] density matrix*/, const bool isstress, ModuleBase::matrix& svnl_dalpha);
+    
+    ///compute Pulay term of the force
+    void cal_f_delta_pulay(const ModuleBase::matrix& dm/**< [in] density matrix*/);
+    void cal_f_delta_k_pulay(const std::vector<ModuleBase::ComplexMatrix>& dm/**<[in] density matrix*/);
+
+    ///calculate tr(\rho V_delta)
+    void cal_e_delta_band(const std::vector<ModuleBase::matrix>& dm/**<[in] density matrix*/);
+    void cal_e_delta_band_k(const std::vector<ModuleBase::ComplexMatrix>& dm/**<[in] density matrix*/);
+
+//============================
+//DeePKS Part 3
+//subroutines that deals with io as well as interface with libtorch
+//realized in LCAO_descriptor_io.cpp
+//============================
+
+    ///calculate partial of energy correction to descriptors
     void cal_gedm(const ModuleBase::matrix& dm/**< [in] density matrix*/);	//need to load model in this step
 
-    ///calculate \f$\frac{d\mathbf{d^I_{nl}}}{dX}=\frac{d\mathbf{d^I_{nl}}}{dD^I_{nlmm'}}*\frac{dD^I_{nlmm'}}{dX}\f$
+    ///calculates gradient of descriptors w.r.t atomic positions
     ///----------------------------------------------------
     ///m, n: 2*l+1
     ///v: eigenvalues of dm , 2*l+1
@@ -90,33 +124,6 @@ public:
     ///gvdm*gdmx->gvx
     ///----------------------------------------------------
     void cal_gvx(const ModuleBase::matrix &dm);
-    
-    ///calculate \f$\sum_{I}\sum_{nlmm'}\langle\phi_\mu|\alpha^I_{nlm}\rangle{\frac{dE}{dD^I_{nlmm'}}}\langle\alpha^I_{nlm'}|\phi_\nu\rangle\f$ (for gamma_only)
-    void build_v_delta_alpha(const bool& cal_deri/**< [in] 0 for 3-center intergration, 1 for its derivation*/);
-    void build_v_delta_alpha_new(const bool& cal_deri/**< [in] 0 for 3-center intergration, 1 for its derivation*/);
-    
-    ///calculate \f$\sum_{I}\sum_{nlmm'}\langle\phi_\mu|\alpha^I_{nlm}\rangle{\frac{dE}{dD^I_{nlmm'}}}\langle\alpha^I_{nlm'}|\phi_\nu\rangle\f$ (for multi-k)
-    void build_v_delta_mu(const bool &cal_deri/**< [in] 0 for 3-center intergration, 1 for its derivation*/);
-    
-    ///compute \f$H_{\delta, \mu\nu} = \langle\phi_\mu|V_\delta|\phi_\nu\rangle\f$ 
-    void cal_v_delta(const ModuleBase::matrix& dm/**< [in] density matrix*/);
-    
-    ///add \f$H_{\delta, \mu\nu}\f$ to the Hamiltonian matrix
-    void add_v_delta(void);
-    void add_v_delta_k(const int &ik);
-
-    ///compute Hellmann-Feynman term of the force contribution of \f$E_\delta\f$
-    void cal_f_delta_hf(const ModuleBase::matrix& dm/**< [in] density matrix*/);
-    void cal_f_delta_hf_new(const ModuleBase::matrix& dm/**< [in] density matrix*/, const bool isstress, ModuleBase::matrix& svnl_dalpha);
-    void cal_f_delta_hf_k_new(const std::vector<ModuleBase::ComplexMatrix>& dm/**<[in] density matrix*/, const bool isstress, ModuleBase::matrix& svnl_dalpha);
-    
-    ///compute Pulay  term of the force contribution of \f$E_\delta\f$
-    void cal_f_delta_pulay(const ModuleBase::matrix& dm/**< [in] density matrix*/);
-    void cal_f_delta_k_pulay(const std::vector<ModuleBase::ComplexMatrix>& dm/**<[in] density matrix*/);
-    
-    ///compute the force contribution of \f$E_\delta\f$
-    void cal_f_delta(const ModuleBase::matrix& dm/**< [in] density matrix*/);
-
 
     ///print descriptors based on LCAO basis
     void print_descriptor(void);
@@ -128,7 +135,7 @@ public:
     void print_F_delta(const std::string &fname/**< [in] the name of output file*/);
 
 	///----------------------------------------------------------------------
-	///The following 3 functions save the `[dm_eig], [e_base], [f_base]`
+	///The following 4 functions save the `[dm_eig], [e_base], [f_base], [grad_vx]`
 	///of current configuration as `.npy` file, when `deepks_scf = 1`.
 	///After a full group of consfigurations are calculated,
     ///we need a python script to `load` and `torch.cat` these `.npy` files,
@@ -143,9 +150,26 @@ public:
 	void save_npy_f(const ModuleBase::matrix &fbase/**<[in] \f$F_{base}\f$ or \f$F_{tot}\f$, in Ry/Bohr*/, const std::string &f_file);
     void save_npy_gvx(void);
 
-    ///calculate \f$tr(\rho H_\delta), \rho = \sum_i{c_{i, \mu}c_{i,\nu}} \f$ (for gamma_only)
-    void cal_e_delta_band(const std::vector<ModuleBase::matrix>& dm/**<[in] density matrix*/);
-    void cal_e_delta_band_k(const std::vector<ModuleBase::ComplexMatrix>& dm/**<[in] density matrix*/);
+//==============
+//obsolete subroutines, in LCAO_descriptor_old.cpp
+//==============
+
+    //calculates descriptors for atoms
+    void cal_dm_as_descriptor(const ModuleBase::matrix& dm/**< [in] density matrix*/); // mohan add 2021-08-04
+
+    ///calculate \f$\sum_{I}\sum_{nlmm'}\langle\phi_\mu|\alpha^I_{nlm}\rangle{\frac{dE}{dD^I_{nlmm'}}}\langle\alpha^I_{nlm'}|\phi_\nu\rangle\f$ (for gamma_only)
+    void build_v_delta_alpha(const bool& cal_deri/**< [in] 0 for 3-center intergration, 1 for its derivation*/);    
+    ///calculate \f$\sum_{I}\sum_{nlmm'}\langle\phi_\mu|\alpha^I_{nlm}\rangle{\frac{dE}{dD^I_{nlmm'}}}\langle\alpha^I_{nlm'}|\phi_\nu\rangle\f$ (for multi-k)
+    void build_v_delta_mu(const bool &cal_deri/**< [in] 0 for 3-center intergration, 1 for its derivation*/);
+    
+    ///compute \f$H_{\delta, \mu\nu} = \langle\phi_\mu|V_\delta|\phi_\nu\rangle\f$ 
+    void cal_v_delta(const ModuleBase::matrix& dm/**< [in] density matrix*/);
+
+    ///compute Hellmann-Feynman term of the force contribution of \f$E_\delta\f$
+    void cal_f_delta_hf(const ModuleBase::matrix& dm/**< [in] density matrix*/);
+    ///compute the force contribution of \f$E_\delta\f$
+    void cal_f_delta(const ModuleBase::matrix& dm/**< [in] density matrix*/);
+
 //-------------------
 // public variables
 //-------------------
@@ -240,14 +264,23 @@ private:
 //-------------------
 private:
 
-	void init_index(void);	// index of descriptor in all atoms
+//===============================
+//DeePKS Part 1
+//deals with generation of descriptors as well as labels
+//realized in LCAO_descriptor.cpp
+//===============================
 
+// arrange index of descriptor in all atoms
+	void init_index(void);
+
+//save the overlap <alpha|chi> into array
 	void set_S_mu_alpha(
 		const int &iw1_all,
 		const int& inl,
 		const int& im,
 		const double& v);
 
+//for checking purpose, print the projected density matrices
     void print_projected_DM(
 		std::ofstream &ofs,
 		ModuleBase::ComplexMatrix &des,
@@ -256,6 +289,7 @@ private:
 		const int &l,
 		const int& n);
 
+//save the d/dX<alpha|chi> into array
 	void set_DS_mu_alpha(
 		const int& iw1_all,
 		const int& inl,
@@ -264,18 +298,39 @@ private:
 		const double& vy,
 		const double& vz);
 
+//converts density matrix to format used in deepks subroutines
+	void getdm_double(const ModuleBase::matrix& dm);
+
+//calculate the gradient of pdm with regard to atomic positions
+//d/dX D_{Inl,mm'}
 	void init_gdmx(void);
-    void load_model(const std::string& model_file);
-    
-    void cal_gvdm();    //called when force=1, precondition of cal_gvx
     void cal_gdmx(const ModuleBase::matrix& dm);	//dD/dX, precondition of cal_gvx
 	void del_gdmx(void);
 
-	void getdm_double(const ModuleBase::matrix& dm);
+//===============================
+//DeePKS Part 2
+//deals with application of correction dV to Hamiltonian and force
+//realized in LCAO_descriptor_dV.cpp
+//===============================
 
+    //load the trained neural network model
+    void load_model(const std::string& model_file);
+
+//============================
+//DeePKS Part 3
+//subroutines that deals with io as well as interface with libtorch
+//realized in LCAO_descriptor_io.cpp
+//============================    
+    //partial of descriptors w.r.t projected dm
+    //called when force=1, precondition of cal_gvx
+    void cal_gvdm();    
+
+    //converts descriptor from projected density matrices
+    //and converts to libtorch format
 	void cal_descriptor_tensor(void);
 
 };
+
 namespace GlobalC
 {
 extern LCAO_Descriptor ld;
